requirements.txt
# Firebase / Google Cloud
google-cloud-firestore==2.16.1
google-auth-oauthlib==1.2.0
google-auth==2.25.2
google-cloud-core==2.3.3

# Streamlit
streamlit==1.28.1

# Utilities (optional)
python-dotenv==1.0.0


firebase-admin

pyrebase4

streamlit-autorefresh

# read the files

tabulate